requirements.txt
Django==3.1.4
braintree
django-filter==2.4.0
Pillow==8.1.0
djangorestframework==3.12.2
